chore: move `getBaseURI` to `ui-client` package
