+++ packages/gazzodown/src/elements/LinkSpan.tsx
@@ -1,4 +1,5 @@
 import type * as MessageParser from '@rocket.chat/message-parser';
+import { getBaseURI, isExternal } from '@rocket.chat/ui-client';
 import { useTranslation } from '@rocket.chat/ui-contexts';
 import { ReactElement, useMemo } from 'react';
 
@@ -6,27 +7,6 @@ import BoldSpan from './BoldSpan';
 import ItalicSpan from './ItalicSpan';
 import PlainSpan from './PlainSpan';
 import StrikeSpan from './StrikeSpan';
-
-const getBaseURI = (): string => {
-	if (document.baseURI) {
-		return document.baseURI;
-	}
-
-	// Should be exactly one tag:
-	//   https://developer.mozilla.org/en-US/docs/Web/HTML/Element/base
-	const base = document.getElementsByTagName('base');
-
-	// Return location from BASE tag.
-	if (base.length > 0) {
-		return base[0].href;
-	}
-
-	// Else use implementation of documentURI:
-	//   http://www.w3.org/TR/DOM-Level-3-Core/core.html#Node3-baseURI
-	return document.URL;
-};
-
-const isExternal = (href: string): boolean => href.indexOf(getBaseURI()) !== 0;
 
 type LinkSpanProps = {
 	href: string;
