+++ packages/ui-client/src/helpers/index.ts
@@ -1,0 +1,1 @@
+export * from './getBaseURI';
