+++ packages/ui-client/src/index.ts
@@ -2,3 +2,4 @@ export * from './components';
 export * from './hooks/useFeaturePreview';
 export * from './hooks/useFeaturePreviewList';
 export * from './hooks/useDocumentTitle';
+export * from './helpers';
